Forming generic Job Adapter
